IC4C-159: tcp_receive_packet was not handling connector_callback_error.
	- tcp_receive_packet now returns a connector_unavailable, which
can be handled in connector_step().
	- On connector_tcp_open.h, changed the close_status for one less
harmful. In this case, this was causing the abort when the above
connector_unavailable was received. With connector_close-status_device_error
a retry can be done manually and the thread is not exited.
	- Fixes: https://jira.digi.com/browse/IC4C-159
(cherry picked from commit 503489a4f2b30d3714b2c92437dd2ca5b2fb1465)