IC4C-287: Send all the CVS entries in all DataPoints even if they are not necessary.
	- Function dp_update_state() was always putting a comma (',') at the end of each entry.
Then, if it was the last one, dp_info->data.csv.last_entry_ptr was being used to replace it with a
'\n' if there were more DPs to add to the packet; else it returned a 0, marking the end of the packet
only if the function was called again. When {bytes_remaining}} was exactly "1" in the last entry
(description), then it got out of the loop without calling dp_update_state() again, therefore, the
state machine flow was not valid.
	- I really hope this doesn't fuck something else up... Go ahead, punk... Make my day... ¬_¬
	- Code is not optimized. Related: https://jira.digi.com/browse/IC4C-288
	- Fixes: https://jira.digi.com/browse/IC4C-287