rci: Ensure collection opened on set even if no elements touched
The Configuration Service attempts to use the complete attribute
and mark unchanged dictionary entries as still wanted by sending
the element with no fields. As we currently inform the implementation
layer that a entry should not be removed by calling the open callback
we must make sure this happens even when no elements are being set.

IC4C-674