CC: impliment oneshot command
Implimented the oneshot command for the cc. A callback is triggered
during the request stage that stores the command as is in a temporary
file. During the request stage the contents of the file are directly
given to the cli which sends its output to another file. Files are used
to prevent any weirdness with special characters.

These changes to how sessions are managed solve the error found in
DAL-4412.

Also rearranged msg_service_type_need_data and
msg_service_type_have_data in data_service_device_request_callback in
private/connector_data_service.h to match processing order.

DAL-4273, DAL-4412