rci_binary_callback: Prevent use of uninitialised list depth
When running multiple RCI set requests,  the rci data is allocated
and freed.  Most values inside rci_internal_data are initialised
appropriately before use,  but rci->shared.callback_data.list.depth
is not.  It is used by the should_run_callback() macro in all states
including session/action_start,  yet is not initialised until
group/list_start which happens later.

This meant that the list depth checks in should_run_callback() were
making decisions on uninitialised data and not always calling required
setup code.  Because this setup code was not run,  in later states we
would crash because values were not setup as required.

In this case we would assert because a DAL libconfig object (more invalid data)
was not an object.

Rather than rework the questionable logic in should_run_callback(),  I have
opted to initialise calback_data.list.depth next to some related code that
was already being used to initialise data in arguably the wrong place,

This fixes the crash with the minimal of change and ensures all
required call backs are run.

DAL-4593