Remove invalid pointer cast
Accessing an lvalue through an incompatible type such as when casting a pointer
to a pointer of a different type and dereferencing the casted pointer is
undefined behavior. On GCC 11.2 this caused a warning about 'value' being used
while possibly undefined. Removing the pointer cast and instead using memcpy
causes the warning to disapper.