connector_msg.h: fix file system not working if compression buffers size is equal or bigger than available window
If 'MSG_MAX_RECV_PACKET_SIZE' is bigger or equal than the available window, the
file system feature wrongly waits for an ack that is never received and fails
with a timeout (see function 'msg_process_send_data()')

We have seen this after increasing the value of 'MSG_MAX_RECV_PACKET_SIZE' to
10K to improve the firmware download speed
(https://onedigi.atlassian.net/browse/DEL-8750)

In this case and with an available window of also 10K, function
'msg_process_send_data()' states an ack is pending and the connector waits for
an ack that will never be received.

This commit uses 'MSG_MAX_SEND_PACKET_SIZE' for the size of buffers.

https://onedigi.atlassian.net/browse/DEL-8832

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>